core: remove neointerops-related files
And move their content to systeminterops-related files.
